Move extension to shared module
